Fixed Code Inspection warnings

+++ src/main/java/com/arcao/geocaching/api/impl/live_geocaching_api/filter/FavoritePointsFilter.java
@@ -20,7 +20,7 @@ public class FavoritePointsFilter implements Filter {
   }
 
 	public boolean isValid() {
-		if (min != null && max != null && min.intValue() > max.intValue())
+		if (min != null && max != null && min > max)
 			return false;
 		
 	  return (min != null || max != null);
